backend/pyproject.toml
[project]
name = "spotify-oauth-backend"
version = "0.1.0"
description = "FastAPI backend for Mr. DJ - AI-powered Spotify playlist generator"
requires-python = ">=3.12.1"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "spotipy>=2.25.1",
    "langgraph>=0.6.6",
    "langchain-openai>=0.3.32",
    "langchain-core>=0.3.0",
    "langsmith>=0.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "langchain-community>=0.4.1",
    "langchain-tavily>=0.2.13",
    "langmem>=0.0.30",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
